+++ deepctr/layers/interaction.py
@@ -1431,9 +1431,9 @@ class FEFMLayer(Layer):
          https://arxiv.org/pdf/2009.09931.pdf
     """
 
-    def __init__(self, num_fields, embedding_size, regularizer, **kwargs):
-        self.num_fields = num_fields
-        self.embedding_size = embedding_size
+    def __init__(self, regularizer, **kwargs):
+        #self.num_fields = num_fields
+        #self.embedding_size = embedding_size
         self.regularizer = regularizer
         super(FEFMLayer, self).__init__(**kwargs)
 
@@ -1442,16 +1442,16 @@ class FEFMLayer(Layer):
             raise ValueError("Unexpected inputs dimensions % d,\
                                 expect to be 3 dimensions" % (len(input_shape)))
 
-        if input_shape[1] != self.num_fields:
-            raise ValueError("Mismatch in number of fields {} and \
-                    concatenated embeddings dims {}".format(self.num_fields, input_shape[2]))
-
+        # if input_shape[1] != self.num_fields:
+        #     raise ValueError("Mismatch in number of fields {} and \
+        #             concatenated embeddings dims {}".format(self.num_fields, input_shape[2]))
+        self.num_fields  = int(input_shape[1])
         self.field_embeddings = {}
-
+        embedding_size = int(input_shape[2])
         for fi, fj in itertools.combinations(range(self.num_fields), 2):
             field_pair_id = str(fi) + "-" + str(fj)
             self.field_embeddings[field_pair_id] = self.add_weight(name='field_embeddings' + field_pair_id,
-                                                                   shape=(self.embedding_size, self.embedding_size),
+                                                                   shape=(embedding_size, embedding_size),
                                                                    initializer=TruncatedNormal(),
                                                                    regularizer=l2(self.regularizer),
                                                                    trainable=True)
@@ -1464,9 +1464,9 @@ class FEFMLayer(Layer):
                 "Unexpected inputs dimensions %d, expect to be 3 dimensions"
                 % (K.ndim(inputs)))
 
-        if inputs.shape[1] != self.num_fields:
-            raise ValueError("Mismatch in number of fields {} and \
-                    concatenated embeddings dims {}".format(self.num_fields, inputs.shape[1]))
+        # if inputs.shape[1] != self.num_fields:
+        #     raise ValueError("Mismatch in number of fields {} and \
+        #             concatenated embeddings dims {}".format(self.num_fields, inputs.shape[1]))
 
         pairwise_inner_prods = []
         for fi, fj in itertools.combinations(range(self.num_fields), 2):
@@ -1489,8 +1489,8 @@ class FEFMLayer(Layer):
     def get_config(self):
         config = super(FEFMLayer, self).get_config().copy()
         config.update({
-            'num_fields': self.num_fields,
+            #'num_fields': self.num_fields,
             'regularizer': self.regularizer,
-            'embedding_size': self.embedding_size
+            #'embedding_size': self.embedding_size
         })
         return config